Add create shipment button component
